advanced django/django/my_django_project/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.12,<4.0
drf-yasg>=1.21.0,<2.0.0
psycopg2-binary>=2.9,<3.0
gunicorn>=20.1,<21.0
django-cors-headers>=3.10,<4.0
python-dotenv>=0.19,<1.0